Project 4 completed
project 4 was completed in the tomcat directory unlike my other projects and repos. This is a transfer of all files from that tomcat location and onedrive resource files (pdf's and pics for ex). This project is final and is my last submission for my undergraduate degree (including finals).

GK;CO!